Use correct health endpoints
The readiness and liveness probes were previously set to /.
This meant that the probes failed, as the route / needs authentication.

The management-ui already has endpoints for both liveness and readiness,
so this commit points the probes at the correct endpoints.

The liveness probe on the management-ui checks to see if the
management-ui itself is ready to run. If the liveness check fails, then
Kubernetes will make the pod restart.

The readiness probe is a bit more complicated. Kubernetes won't restart
if the readiness probe fails, but it won't direct traffic towards the
pod until it does. The management-ui currently checks to see if it can
connect to all the relevant upstream APIs before returning a successful
readiness check.