Hocs 2097 autoscale replicas (#29)
* HOCS-2097 Autoscale Replicas
The minimum instances for a given target environment is set in the deploy script. Using this value to set the default number of minimum instances for the service.
Also amended the max number and the target percentage, which are incorrect.

* HOCS-2097 Autoscale Replicas
Maximum replicas have been added to the autoscaler.
The value is set in the deploy script along with the minimum replicas.