HOCS-2381: Upgrade eslint to handle typescript
As a result of tslint being out of date and eslint now allowing for
testing of typescript files, this commit brings eslint upto the latest
release.
Previously we weren't running linting on .tsx files, this commit also
also adds linting to .tsx.
Remove the tslint dependencies as they are no longer required.
Add .idea folder to the .eslintignore.
ddee