HOCS-2381: Fix linting and add to build process (#72)
* HOCS-2381: Upgrade eslint to handle typescript

As a result of tslint being out of date and eslint now allowing for
testing of typescript files, this commit brings eslint upto the latest
release.
Previously we weren't running linting on .tsx files, this commit also
also adds linting to .tsx.
Remove the tslint dependencies as they are no longer required.
Add .idea folder to the .eslintignore.
ddee

* HOCS-2381: change logstash application name

Since the forming of MUI the application name for logstash has
been set to hocs-frontend. This commit changes this to be
hocs-management-ui.

* HOCS-2381: first fix for lint issues

As a result of fixing the linting this flagged up a number of issues.
This commit acts as a first fix for lint issues that are easily
resolvable. Tickets will be generated for fixing the more complex
linting issues.

* HOCS-2381 Add linter to pipeline

This commit also removes the build condition on the Docker build, to
ensure we test and lint every time (instead of just on `push`).

* HOCS-2381: Add ticket numbers for todos

Due to turning off specific linting rules that probably should
be active, this commit adds ticket numbers to each of the todo's
that have been added to the code.

* HOCS-2381: correct rebase lint issue

As a result of rebasing of master a linting issue has been
introduced. This commit fixes this issue.

Co-authored-by: Douglas Gardner <douglas@chippy.ch>