HOCS-2644: drone v1 migration (#86)
* HOCS-2644: migrate from drone 0.8 to 1

This change implements the necessary changes to use Drone V1 as V0.8
is going to be deprecated. Changes include:
- Add build, test to pipeline (previously wasn't testing the code, only
ensuring the build was successful
- Removed pull request checks
- Changed to push to build_XXX instead of build-XXX
- Added commit sha tag
- Switched to using KUBE_TOKEN as defined in Drone rather than the
kube repo
- Added acyclic dependency graph for asynchronous build steps
- Made a QA release always release to both QA environments
consolidated build steps where possible
- Ensure production build is build (previously was development)

* HOCS-2644: run npm audit --fix

Run `npm audit --fix`, this audits all the dependencies through the 
application and updates them. This improves the reports from 36 to 10 
to eventually fix.