HOCS-2502: fix-up add new team (#90)
* HOCS-2502: change page title

At present the Add Team page makes the user assume that they
can add a team. This feature is currently limited to adding a DCU
drafting team. As a result the page title and what is shown on the
dashboard should reflect this.

* HOCS-2502: change to force all 3 DCU permissions

At present the user selects one of the permissions and the team is
then added. This has the unintended issue that one team will not be
able to access the other 2 DCU areas. This change removes the case
type from the team creation and essentially hard codes the case
types of all 3 DCU areas.

* HOCS-2502: protect backend for add new DCU user

Currently anyone can hit the backend API to create a new DCU
drafting team. This change adds the protect middleware to check
that the user has the Keycloak DCU role.

* HOCS-2502: add hasRole function to state

Currently we have no way of blocking UI components based on user
roles. This change adds a `hasRole` function to the ApplicationState.
This can then be used to 'feature-gate' dependant on the users permissions.

* HOCS-2502: feature-gate adding dcu drafting team

Currently anyone has access to create a DCU drafting team. This
should be limited to only the DCU users, as a result of this change a
user role check is applied that stops the showing of the link to those
that don't have the `DCU` role in keycloak.

* HOCS-2502: whitespace cleanup