HOCS-2743: edit team name (#91)
* HOCS-2743: add updateTeamName middleware

To allow for the updating of the team name we need a middleware 
that will call the info-service to action. This change adds a 
`updateTeamName` function and the necessary tests.

* HOCS-2743: add server routing for updating team name

Add the server route for updating a team at the PUT 
`api/team/{teamId}` endpoint.

* HOCS-2743: add axios call to API 

Add the axios put call to the node server API that passes through the 
team uuid and the new display name to update in info-service.

* HOCS-2743: change team model

Currently the `displayName` in the team model allows for an 
optional string or undefined. This is not the case and we enforce that 
a name is not null. Therefore to male the model more strict this
limitation has been removed.

* HOCS-2743: add editTeam page

Add the editTeam page, state, reducer and actions. This page allows the user to rename a specific team based on the teamId (team uuid) passed in through the params. This then submits through the frontend to the node server and info-service.

* HOCS-2743: add edit team button to teamview

Add the edit team button to the team view page locked behind the RENAME_TEAM role.

* HOCS-2743: protect rename endpoint

Protect the node server endpoint for renaming a team behind the
RENAME_TEAM role.