Remove questionable DNS_PREFIX stuff
We had some variables that would take a prefix and a suffix and join
them together, even though the suffix never changes. Now previous
commits have simplified the script in a way that means the prefix is
declared less often, there's no real reason to keep the prefix and
suffix separate.

This commit changes it to use the FQDN instead, in all references.