HOCS-5552 Update GOVUK Frontend to 4.3.1 (#294)
Updates GOVUK Frontend package to 4.3.1, and makes required changes for affected elements between the versions.

Impacted areas
- Header
The HTML for the header was updated in 4.0.0 for better screen reader support.

- Hint text
Hint text elements were switched to divs from spans in 3.8.0, 4.0.0 introduced a breaking change were the styles were removed for the spans. The hint text spans have been converted to divs.

- Error messages
Error messages now use p tags rather than span

- Change date input from type="number" to inputmode="numeric"
This was a Recommended change in 4.0.0 following research by the Design System team. In 4.1.0 a Recommended change was made to remove the pattern from the element.

- Footer
Fix fallback logo so Chrome will not flag it to screen readers.

Co-authored-by: Douglas Gardner <douglas.gardner@digital.homeoffice.gov.uk>